feat(scoped-elements): work without polyfill if possible
