+++ packages/scoped-elements/package.json
@@ -43,8 +43,10 @@
   ],
   "dependencies": {
     "@lit/reactive-element": "^1.0.0",
-    "@open-wc/dedupe-mixin": "^1.3.0",
-    "@webcomponents/scoped-custom-element-registry": "^0.0.3"
+    "@open-wc/dedupe-mixin": "^1.3.0"
+  },
+  "devDependencies": {
+    "@webcomponents/scoped-custom-element-registry": "^0.0.5"
   },
   "types": "types/index.d.ts"
 }
